Add all ships and lasers in separate files and folders
