+++ src/Entities/EnemyLaser.js
@@ -1,0 +1,10 @@
+import Entity from "../js/Entities"
+
+class EnemyLaser extends Entity {
+    constructor(scene, x, y) {
+      super(scene, x, y, "sprLaserEnemy0");
+      this.body.velocity.y = 200;
+    }
+}
+
+export default EnemyLaser;
