+++ src/Entities/PlayerLaser.js
@@ -1,0 +1,9 @@
+import Entity from "../js/Entities"
+
+class PlayerLaser extends Entity {
+    constructor(scene, x, y) {
+      super(scene, x, y, "sprLaserPlayer");
+      this.body.velocity.y = -200;
+    }
+}
+export default PlayerLaser;
