+++ src/Entities/ScrollingBackground.js
@@ -1,3 +1,4 @@
+import Phaser from "phaser";
 class ScrollingBackground {
     constructor(scene, key, velocityY) {
         this.scene = scene;
