+++ src/js/SceneGameOver.js
@@ -1,3 +1,4 @@
+import Phaser from 'phaser';
 class SceneGameOver extends Phaser.Scene {
     constructor() {
       super({ key: "SceneGameOver" });
