--- src/js/game.js
@@ -1,20 +1,0 @@
-var config = {
-    type: Phaser.WEBGL,
-    width: 480,
-    height: 640,
-    backgroundColor: "black",
-    physics: {
-      default: "arcade",
-      arcade: {
-        gravity: { x: 0, y: 0 }
-      }
-    },
-    scene: [
-        SceneMainMenu,
-        SceneMain,
-        SceneGameOver
-    ],
-    pixelArt: true,
-    roundPixels: true
-};
-var game = new Phaser.Game(config);
